remove unnecessary float lookups from numpy module

+++ astropy/io/misc/yaml.py
@@ -266,7 +266,7 @@ for np_type in [np.int_, np.int, np.intc, np.intp, np.int8, np.int16, np.int32,
                 np.int64, np.uint8, np.uint16, np.uint32, np.uint64]:
     AstropyDumper.add_representer(np_type,
                                  yaml.representer.SafeRepresenter.represent_int)
-for np_type in [np.float_, np.float, np.float16, np.float32, np.float64,
+for np_type in [np.float_, float, np.float16, np.float32, np.float64,
                 np.longdouble]:
     AstropyDumper.add_representer(np_type,
                                  yaml.representer.SafeRepresenter.represent_float)
